feat: init revamp
