+++ Cargo.toml
@@ -1,31 +1,11 @@
 [package]
-name = "Mori"
-version = "0.1.0"
+name = "mori"
+version = "1.0.0"
 edition = "2021"
 
 [dependencies]
-tokio = { version = "1.38.0", features = ["full"] }
-enet = { git = "https://github.com/CLOEI/enet-rs" }
-gtworld-r = { git = "https://github.com/CLOEI/gtworld-r" }
-gtitem-r = { git = "https://github.com/CLOEI/gtitem-r", features = ["serde"] }
-regex = "1.10.5"
-spdlog-rs = "0.3.13"
-ureq = { version = "2.10.0", features = ["cookies"] }
-json = "0.12.4"
+bincode = "1.3.3"
 byteorder = "1.5.0"
 rand = "0.8.5"
-sha2 = "0.10.8"
-md5 = "0.7.0"
-hex = "0.4.3"
-base64 = "0.22.1"
-rust-otp = "2.0.0"
-serde_json = "1.0.120"
-serde = { version = "1.0.204", features = ["derive"] }
-eframe = "0.28.1"
-chromiumoxide = { version = "0.6.0", features = [
-  "tokio-runtime",
-], default-features = false }
-futures = "0.3.30"
-egui_extras = { version = "*", features = ["all_loaders"] }
-image = { version = "0.25", features = ["png"] }
+ureq = "2.10.1"
 urlencoding = "2.1.3"
