--- src/bot/features/auto_tutorial.rs
@@ -1,1 +1,0 @@
-pub fn start() {}
