--- src/gui/mod.rs
@@ -1,4 +1,0 @@
-pub mod add_bot_dialog;
-pub mod bot_menu;
-pub mod item_database;
-pub mod navbar;
