+++ src/types/epacket_type.rs
@@ -1,0 +1,14 @@
+#[repr(u32)]
+#[derive(Debug)]
+enum EPacketType {
+    NetMessageUnknown,
+    NetMessageServerHello,
+    NetMessageGenericText,
+    NetMessageGameMessage,
+    NetMessageGamePacket,
+    NetMessageError,
+    NetMessageTrack,
+    NetMessageClientLogRequest,
+    NetMessageClientLogResponse,
+    NetMessageMax,
+}
