+++ src/types/etank_packet_type.rs
@@ -1,0 +1,44 @@
+#[repr(u8)]
+#[derive(Debug)]
+pub enum ETankPacketType {
+    NetGamePacketState,
+    NetGamePacketCallFunction,
+    NetGamePacketUpdateStatus,
+    NetGamePacketTileChangeRequest,
+    NetGamePacketSendMapData,
+    NetGamePacketSendTileUpdateData,
+    NetGamePacketSendTileUpdateDataMultiple,
+    NetGamePacketTileActivateRequest,
+    NetGamePacketTileApplyDamage,
+    NetGamePacketSendInventoryState,
+    NetGamePacketItemActivateRequest,
+    NetGamePacketItemActivateObjectRequest,
+    NetGamePacketSendTileTreeState,
+    NetGamePacketModifyItemInventory,
+    NetGamePacketItemChangeObject,
+    NetGamePacketSendLock,
+    NetGamePacketSendItemDatabaseData,
+    NetGamePacketSendParticleEffect,
+    NetGamePacketSetIconState,
+    NetGamePacketItemEffect,
+    NetGamePacketSetCharacterState,
+    NetGamePacketPingReply,
+    NetGamePacketPingRequest,
+    NetGamePacketGotPunched,
+    NetGamePacketAppCheckResponse,
+    NetGamePacketAppIntegrityFail,
+    NetGamePacketDisconnect,
+    NetGamePacketBattleJoin,
+    NetGamePacketBattleEvent,
+    NetGamePacketUseDoor,
+    NetGamePacketSendParental,
+    NetGamePacketGoneFishin,
+    NetGamePacketSteam,
+    NetGamePacketPetBattle,
+    NetGamePacketNpc,
+    NetGamePacketSpecial,
+    NetGamePacketSendParticleEffectV2,
+    NetGamePacketActiveArrowToItem,
+    NetGamePacketSelectTileIndex,
+    NetGamePacketSendPlayerTributeData,
+}
