+++ src/types/vector.rs
@@ -1,0 +1,5 @@
+#[derive(Default, Debug)]
+pub struct Vector2 {
+    pub x: f32,
+    pub y: f32,
+}
